+++ src/mobilityPackagePriceEstimation/java/org/opentripplanner/estimator/hasura_client/mappers/VehicleStateSnapshotMapper.java
@@ -6,22 +6,15 @@ import org.opentripplanner.hasura_client.mappers.HasuraToOTPMapper;
 import org.opentripplanner.hasura_client.mappers.VehiclePositionsMapper;
 import org.opentripplanner.routing.core.vehicle_sharing.Provider;
 import org.opentripplanner.routing.core.vehicle_sharing.VehicleDescription;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 
 public class VehicleStateSnapshotMapper extends HasuraToOTPMapper<VehicleStateSnapshotHasuraObject, VehicleDescription> {
-
-    private static final Logger LOG = LoggerFactory.getLogger(VehicleStateSnapshotMapper.class);
 
     private Map<Integer, Provider> vehicleProviders;
 
     public VehicleStateSnapshotMapper(Map<Integer, Provider> vehicleProviders) {
-        this.vehicleProviders = vehicleProviders;
+        this.vehicleProviders = Objects.requireNonNullElse(vehicleProviders, Collections.emptyMap());
     }
 
     @Override
@@ -29,7 +22,8 @@ public class VehicleStateSnapshotMapper extends HasuraToOTPMapper<VehicleStateSn
         VehiclePositionsMapper originalVehicleMapper = new VehiclePositionsMapper();
         Provider vehicleProvider = vehicleProviders.get(vehicleStateSnapshotHasuraObject.getProviderId());
         if (Objects.nonNull(vehicleProvider)) {
-            //TODO: this seems to be a little bit ugly workaround to avoid copying the original mapping method code here...
+            /*This is a little bit ugly workaround to automatically incorporate all future modifications of the original
+            vehicle mapping method code here, without changing the existing vehicle mapper interface*/
             Vehicle stateSnapshotAsVehicle = vehicleStateSnapshotHasuraObject.toVehicle(vehicleProvider);
             List<VehicleDescription> mappedVehicles = originalVehicleMapper.map(Arrays.asList(stateSnapshotAsVehicle));
             return mappedVehicles.get(0);
